+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added try-excepts and timeouts to the different tasks in the Overwatcher shutdown routine to ensure that the dome closure is always attempted.
 * Move `DomeHelper.startup()` to `Overwatcher.startup()` and clean up observer code.
 * Retry emitting Overwatcher heartbeats.
+* Add additional checks to prevent concurrent attempts to open/close the dome.
 
 ### 🔧 Fixed